refactor: `square` 아이콘의 이름을 `rect`로 변경
